requirements.txt
tensorflow>=2.12.0
flask>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
Werkzeug>=3.0.0
